Dummy commit
This was an attempt to fix a problem with the GitHub workflow. The
worflow fails on GitHub although it passes locally. I had a linting
problem when I initially pushed. I fixed this and force-pushed. I'm
afraid it may be caching the pre-force-push code and running the
workflow against that. Hoping these commits will invalidate that cache,
if that is in fact the issue.